Move game classes into correct packages
