+++ chat.llamas.cardlogic.game.api/src/main/java/chat/llamas/cardlogic/game/api/RoundEventManagerInterface.java
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package chat.llamas.cardlogic;
+package chat.llamas.cardlogic.game.api;
 
 import chat.llamas.cardlogic.domain.commands.BetCommand;
 import chat.llamas.cardlogic.domain.commands.CallCommand;
